requirements.txt
python-dotenv==1.2.1
python-docx==1.2.0
requests==2.32.5
pandas==2.3.3
